working way through tests, made minor change to test data
