+++ src/main/java/gov/usgs/cida/wqp/util/HttpConstants.java
@@ -27,6 +27,7 @@ public class HttpConstants {
 	public static final String SIMPLE_STATION_ENDPOINT = "/" + ENDPOINT_SIMPLE_STATION + ENDPOINT_SEARCH;
 	public static final String SUMMARY_ORGANIZATION_ENDPOINT = "/" + ENDPOINT_SUMMARY + ENDPOINT_SUMMARY_ORGANIZATION + ENDPOINT_SEARCH;
 	public static final String SUMMARY_MONITORING_LOCATION_ENDPOINT = "/" + ENDPOINT_SUMMARY + ENDPOINT_MONITORING_LOCATION + ENDPOINT_SEARCH;
+	public static final String PERIOD_OF_RECORD_ENDPOINT = "/" + ENDPOINT_SUMMARY + ENDPOINT_MONITORING_LOCATION + ENDPOINT_SEARCH;
 	public static final String RES_DETECT_QNT_LMT_SEARCH_ENDPOINT = "/" + ENDPOINT_RES_DETECT_QNT_LMT + ENDPOINT_SEARCH;
 	public static final String PROJECT_SEARCH_ENDPOINT = '/' + ENDPOINT_PROJECT + ENDPOINT_SEARCH;
 	public static final String PROJECT_MONITORING_LOCATION_WEIGHTING_SEARCH_ENDPOINT = '/' + ENDPOINT_PROJECT_MONITORING_LOCATION_WEIGHTING + ENDPOINT_SEARCH;
