+++ src/main/resources/logback-spring.xml
@@ -48,15 +48,15 @@
     <!-- end from org/springframework/boot/logging/logback/file-appender.xml -->
 
 
-    <root level="INFO">
+    <root level="DEBUG">
         <appender-ref ref="CONSOLE" />
         <appender-ref ref="FILE" />
     </root>
 
     <logger name="gov.usgs" level="INFO" />
     <logger name="org.springframework" level="INFO" />
-    <logger name="org.mybatis" level="INFO" />
-    <logger name="java.sql" level="INFO" />
+    <logger name="org.mybatis" level="DEBUG" />
+    <logger name="java.sql" level="DEBUG" />
     <logger name="springfox" level="INFO" />
 
 </configuration>
