+++ src/test/resources/logback-spring.xml
@@ -9,14 +9,14 @@
             <pattern>${CONSOLE_LOG_PATTERN}</pattern>
         </encoder>
     </appender>
-    <root level="INFO">
+    <root level="DEBUG">
         <appender-ref ref="CONSOLE" />
     </root>
 
     <logger name="gov.usgs" level="INFO" />
     <logger name="org.springframework" level="INFO" />
-    <logger name="org.mybatis" level="INFO" />
-    <logger name="java.sql" level="INFO" />
+    <logger name="org.mybatis" level="DEBUG" />
+    <logger name="java.sql" level="DEBUG" />
     <logger name="springfox" level="INFO" />
     <logger name="org.dbunit" level="INFO" />
