+++ src/test/resources/testData/clearAll.xml
@@ -1,7 +1,7 @@
 <?xml version='1.0' encoding='UTF-8'?>
 <dataset>
 	<county/>
-	<state/>
+	<state/>	
 	<project_dim/>
 	<station/>
 	<station_sum/>
@@ -12,4 +12,5 @@
 	<web_service_log/>
 	<prj_ml_weighting/>
 	<organization_sum/>
+	<ml_grouping/>
 </dataset>
